Working function of addition new category by a logged user with using Pagination class

+++ src/main/java/pl/project/englishwordswebapp/service/CurrentUser.java
@@ -11,20 +11,20 @@ import org.springframework.web.context.annotation.SessionScope;
 @Scope(scopeName= WebApplicationContext.SCOPE_SESSION, proxyMode= ScopedProxyMode.TARGET_CLASS)
 public class CurrentUser {
 
-    private int id;
+    private Long id;
     private boolean logged = false;
     Integer t = null;
 
     public CurrentUser(){};
-    public void setIdAndLog(int id, boolean logged){
+    public void setIdAndLog(Long id, boolean logged){
         this.id = id;
         this.logged = logged;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public boolean getLogged() {
